Вынесен запрос данных для списка заказов пользователя и для ленты
